START: refactor/extension of spfy api to support group comparisons
